more specific extensions and tests
