+++ src/commonMain/kotlin/csense/kotlin/extensions/primitives/String.kt
@@ -237,30 +237,6 @@ public inline fun String.skipStartsWith(
 }
 //endregion
 
-//region Hex converting
-/**
- * The opposite of [csense.kotlin.extensions.collections.array.toHexString] , so takes a hex string (eg "0x20") and converts it to a byte array of that
- * if any error is found during the "deserialization, null will be returned.
- * @receiver [String]
- * @return [ShortArray]? since [UByte]s are still experimental, [Short]s are used to make sure the size "is ok"
- */
-//TODO specific extension
-public inline fun String.fromHexStringToByteArray(): ShortArray? {
-    //strip prefix iff asked to
-    if (length.isOdd || isEmpty()) {
-        return null
-    }
-    //we have the hex prefix iff it starts with "0x". strip that iff necessary
-    val string = skipStartsWith("0x", true)
-    val result = ShortArray(string.length / 2)
-    GenericCollectionExtensions.forEach2Indexed(string.length, string::get) { index: Int, first: Char, second: Char ->
-        val shortValue = hexCharsToValue(first, second) ?: return@fromHexStringToByteArray null
-        result[index / 2] = shortValue
-    }
-    return result
-}
-//endregion
-
 
 //region Does not starts with
